Completing bootstrap token acquisition part of TestBootstrapScenario_IntTest()
