--- integrationtest/bootstrap/src/test/java/soap/integrationAndInteroperability/WSC_Java_WSP_dotNET.java
@@ -1,14 +1,0 @@
-package soap.integrationAndInteroperability;
-import org.junit.Assert;
-import org.junit.Test;
-import org.openqa.selenium.By;
-
-import java.util.concurrent.TimeUnit;
-
-public class WSC_Java_WSP_dotNET extends Setup {
-
-    @Test
-    public void java_dotNet_SOAP_Bootstrap() throws InterruptedException {
-    }
-
-}
